requirements.txt
dspy-ai
manim
moviepy
openai
google-generativeai
pytest
gTTS